feat: stream responses

+++ frontend/components/chat_page.py
@@ -5,7 +5,8 @@ from loguru import logger
 
 from frontend.api import APIClient
 from frontend.components import *
-from frontend.datatypes import MessagePair
+from frontend.datatypes import MessagePair, Step
+from frontend.utils.constants import NEW_CHAT
 from frontend.utils.icons import AVATARS
 
 
@@ -23,13 +24,13 @@ class ChatPage:
         self.logger = logger.bind(classname=self.__class__.__name__)
 
     def _create_thread_and_register(self, title: str) -> bool:
-        """Create a thread for this chat page and add itself to the chat pages list
+        """Create a thread for this chat page and add itself to the chat pages list.
 
         Args:
-            title (str): The thread title
+            title (str): The thread title.
 
         Returns:
-            bool: Whether the thread creation was successful
+            bool: Whether the thread creation was successful.
         """
         thread = self.api.create_thread(
             access_token=st.session_state["access_token"],
@@ -47,11 +48,11 @@ class ChatPage:
             return False
 
     def _handle_click_feedback(self, feedback_id: str, show_comments_id: str):
-        """Update the feedback buttons state and the comments text input flag on session state
+        """Update the feedback buttons state and the comments text input flag on session state.
 
         Args:
-            feedback_id (str): The feedback button identifier
-            show_comments_id (str): The comments text input flag identifier
+            feedback_id (str): The feedback button identifier.
+            show_comments_id (str): The comments text input flag identifier.
         """
         feedback = st.session_state[feedback_id]
 
@@ -80,15 +81,15 @@ class ChatPage:
         feedback_id: str,
         message_pair_id: uuid.UUID,
         show_comments_id: str,
-        comments: str
+        comments: str | None
     ):
-        """Handle feedback sending
+        """Handle feedback sending.
 
         Args:
-            feedback_id (str): The feedback id
-            message_pair_id (UUID): The message pair id
-            show_comments_id (str): The show_comments flag id
-            comments (str): The comments
+            feedback_id (str): The feedback id.
+            message_pair_id (UUID): The message pair id.
+            show_comments_id (str): The show_comments flag id.
+            comments (str | None): The comments.
         """
         rating = st.session_state[feedback_id]
         access_token = st.session_state["access_token"]
@@ -103,24 +104,17 @@ class ChatPage:
 
         st.session_state[show_comments_id] = False
 
-    def _toggle_flag(self, flag_id: str):
-        """Toggle a flag on session state
-
-        Args:
-            flag_id (str): The flag identifier
-        """
-        st.session_state[flag_id] = not st.session_state[flag_id]
-
     def _render_message_buttons(self, message_pair: MessagePair):
-        """Render the code-showing button and all the feedback related widgets on assistant's messages
+        """Render the code-showing button and all the feedback related widgets on assistant's messages.
 
         Args:
             message_pair (MessagePair):
                 A MessagePair object containing:
-                    - id: unique identifier
-                    - user_message: user message
-                    - assistant_message: assistant message
-                    - generated_queries: generated sql queries
+                    - id: unique identifier.
+                    - user_message: user message.
+                    - assistant_message: assistant message.
+                    - error_message: error message.
+                    - generated_queries: generated sql queries.
         """
         # Placeholder for showing the generated SQL queries
         code_placeholder = st.empty()
@@ -140,7 +134,7 @@ class ChatPage:
 
         # Checks if the model is still answering the last question.
         # If yes, all message buttons and comments inputs will be disabled
-        waiting_for_answer = st.session_state[self.waiting_key]
+        waiting_for_answer = st.session_state[self.page_id][self.waiting_key]
 
         # Unique flag and button key for code-showing
         show_code_id = f"show_code_{message_pair.id}"
@@ -175,7 +169,7 @@ class ChatPage:
                 st.button(
                     " ",
                     key=show_code_btn_id,
-                    on_click=self._toggle_flag,
+                    on_click=toggle_flag,
                     args=(show_code_id,),
                     disabled=waiting_for_answer
                 )
@@ -222,30 +216,36 @@ class ChatPage:
                 disabled=waiting_for_answer
             )
 
-    def _handle_user_interaction(self):
-        """Disable all chat message buttons, comments inputs and the chat input while
-        the model is answering a question and enable the chat deletion button rendering
-        """
-        st.session_state[self.page_id][self.delete_btn_key] = False
-        st.session_state[self.waiting_key] = True
-
-    def _delete_page(self):
-        deleted = self.api.delete_thread(
-            access_token=st.session_state["access_token"],
-            thread_id=self.thread_id
-        )
-
-        if deleted:
-            chat_pages: list[ChatPage] = st.session_state["chat_pages"]
-            for i, chat_page in enumerate(chat_pages):
-                if chat_page.thread_id == self.thread_id:
-                    chat_pages.pop(i)
-                    break
-        else:
-            show_error_popup("Não foi possível excluir a conversa.")
-
     def _render_delete_button(self):
-        """Render the chat reset and download buttons"""
+        """Render the chat deletion button."""
+        @st.dialog("Excluir conversa")
+        def show_delete_chat_modal():
+            st.markdown("") # just for spacing
+            st.text("Tem certeza que deseja excluir esta conversa permanentemente?")
+
+            col1, col2, _ = st.columns([1.1, 2.1, 2])
+
+            if col1.button("Cancelar"):
+                st.rerun()
+
+            if col2.button("Sim, excluir", type="primary"):
+                deleted = self.api.delete_thread(
+                    access_token=st.session_state["access_token"],
+                    thread_id=self.thread_id
+                )
+
+                if not deleted:
+                    st.error("Não foi possível excluir a conversa.", icon=":material/error:")
+                    return
+
+                chat_pages: list[ChatPage] = st.session_state["chat_pages"]
+                for i, chat_page in enumerate(chat_pages):
+                    if chat_page.thread_id == self.thread_id:
+                        chat_pages.pop(i)
+                        break
+
+                st.rerun()
+
         page_session_state = st.session_state[self.page_id]
         chat_delete_disabled = page_session_state[self.delete_btn_key]
 
@@ -255,11 +255,18 @@ class ChatPage:
         st.button(
             label="Excluir",
             icon=":material/delete:",
-            on_click=self._delete_page,
+            on_click=show_delete_chat_modal,
         )
 
+    def _handle_user_interaction(self):
+        """Disable all chat message buttons, comments inputs and the chat input while
+        the model is answering a question and enable the chat deletion button rendering.
+        """
+        st.session_state[self.page_id][self.delete_btn_key] = False
+        st.session_state[self.page_id][self.waiting_key] = True
+
     def render(self):
-        """Render the chat page"""
+        """Render the chat page."""
         # Unfortunately, this is necessary for the code-showing button customization
         st.markdown(
             """<link href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined" rel="stylesheet"/>""",
@@ -273,30 +280,30 @@ class ChatPage:
         if self.page_id not in st.session_state:
             st.session_state[self.page_id] = {}
 
-        # Initialize the waiting key, which is used to prevent users from
-        # interacting with the app while the model is answering a question
-        if self.waiting_key not in st.session_state:
-            st.session_state[self.waiting_key] = False
-
         page_session_state = st.session_state[self.page_id]
 
         # Initialize feedback history
         if self.feedbacks_key not in page_session_state:
             page_session_state[self.feedbacks_key] = {}
 
-        # Initialize chat history state and chat deletion flag state
+        # Initialize the waiting key, which is used to prevent users from
+        # interacting with the app while the model is answering a question
+        if self.waiting_key not in page_session_state:
+            page_session_state[self.waiting_key] = False
+
+        # Initialize chat deletion flag
+        if self.delete_btn_key not in page_session_state:
+            page_session_state[self.delete_btn_key] = self.thread_id is None
+
+        # Initialize chat history state
         if self.chat_history_key not in page_session_state:
             if self.thread_id is not None:
                 message_pairs = self.api.get_message_pairs(
                     access_token=st.session_state["access_token"],
                     thread_id=self.thread_id
                 )
-                # If the thread is already created, the deletion button is enabled
-                page_session_state[self.delete_btn_key] = False
             else:
                 message_pairs = []
-                # Otherwise, it is disabled
-                page_session_state[self.delete_btn_key] = True
             page_session_state[self.chat_history_key] = message_pairs or []
 
         chat_history: list[MessagePair] = page_session_state[self.chat_history_key]
@@ -313,14 +320,29 @@ class ChatPage:
 
             with st.chat_message("assistant", avatar=AVATARS["assistant"]):
                 st.empty()
-                st.write(message_pair.assistant_message)
-                self._render_message_buttons(message_pair)
+
+                if message_pair.assistant_message:
+                    label = "Concluído!"
+                    state = "complete"
+                else:
+                    label = "Erro"
+                    state = "error"
+
+                with st.status(label=label, state=state) as status:
+                    for step in message_pair.safe_steps:
+                        st.caption(step.content)
+
+                if message_pair.assistant_message:
+                    st.write(message_pair.assistant_message)
+                    self._render_message_buttons(message_pair)
+                else:
+                    st.error(message_pair.error_message)
 
         # Accept user input
         if user_prompt := st.chat_input(
             "Faça uma pergunta!",
             on_submit=self._handle_user_interaction,
-            disabled=st.session_state[self.waiting_key]
+            disabled=page_session_state[self.waiting_key]
         ):
             # Clear subheader message
             subheader.empty()
@@ -329,34 +351,42 @@ class ChatPage:
             with st.chat_message("user", avatar=AVATARS["user"]):
                 st.write(user_prompt)
 
+            # Create thread only in the first message
+            if (
+                self.thread_id is None and not
+                self._create_thread_and_register(title=user_prompt)
+            ):
+                clear_new_chat_page()
+                return
+
             # Display assistant response in chat message container
             with st.chat_message("assistant", avatar=AVATARS["assistant"]):
-                three_dots_placeholder = st.empty()
+                with st.status("Consultando banco de dados...") as status:
+                    for streaming_status, message in self.api.send_message(
+                        access_token=st.session_state["access_token"],
+                        message=user_prompt,
+                        thread_id=self.thread_id
+                    ):
+                        if streaming_status == "running":
+                            step: Step = message
+                            status.update(label=step.label)
+                            st.caption(step.content)
+                        elif streaming_status == "complete":
+                            message_pair: MessagePair = message
+                            if message_pair.assistant_message:
+                                label = "Concluído!"
+                                state = "complete"
+                            else:
+                                label = "Erro"
+                                state = "error"
+                            status.update(label=label, state=state)
 
-                with three_dots_placeholder:
-                    three_pulsing_dots()
-
-                # Create thread only in the first message
-                if self.thread_id is None:
-                    if not self._create_thread_and_register(title=user_prompt):
-                        # Setting this flag to False doesn't seem necessary here, but it might
-                        # prevent unexpected issues. Keeping it for safety until the underlying
-                        # behavior is better understood.
-                        st.session_state[self.waiting_key] = False
-                        return
-
-                message_pair = self.api.send_message(
-                    access_token=st.session_state["access_token"],
-                    message=user_prompt,
-                    thread_id=self.thread_id
-                )
-
-                three_dots_placeholder.empty()
-
-                _ = st.write_stream(message_pair.stream_words)
-
-                # Render the message buttons
-                self._render_message_buttons(message_pair)
+                if message_pair.assistant_message:
+                    # Render the assistant message and message buttons
+                    _ = st.write_stream(message_pair.stream_words)
+                    self._render_message_buttons(message_pair)
+                else:
+                    st.error(message_pair.error_message)
 
             # Add message pair to chat history
             chat_history.append(message_pair)
@@ -367,15 +397,44 @@ class ChatPage:
         # Render the disclaimer messages
         render_disclaimer()
 
-        if st.session_state[self.waiting_key]:
-            st.session_state[self.waiting_key] = False
-            current_page = st.Page(
-                page=self.render,
-                title=self.title,
-                url_path=str(self.thread_id)
-            )
-            st.switch_page(current_page)
+        if page_session_state[self.waiting_key]:
+            page_session_state[self.waiting_key] = False
+
+            new_chat: ChatPage | None = st.session_state[NEW_CHAT]
+
+            # If this page is a new chat page, switch to it
+            if new_chat and new_chat.page_id == self.page_id:
+                clear_new_chat_page()
+
+                current_page = st.Page(
+                    page=self.render,
+                    title=self.title,
+                    url_path=str(self.thread_id)
+                )
+
+                st.switch_page(current_page)
+            # Otherwise, we're already on it, so just rerun
+            else:
+                st.rerun()
 
 @st.dialog("Erro")
 def show_error_popup(message: str):
+    """Display an error message in a modal.
+
+    Args:
+        message (str): The error message.
+    """
     st.text(message)
+
+def clear_new_chat_page():
+    """Clear new chat page on session state.
+    """
+    st.session_state[NEW_CHAT] = None
+
+def toggle_flag(flag_id: str):
+    """Toggle a flag on session state.
+
+    Args:
+        flag_id (str): The flag identifier.
+    """
+    st.session_state[flag_id] = not st.session_state[flag_id]
